CSN: CSN-asimcore-28
Updating with simcore changes (rev. 2243)

- For well over a year, the port code has been dynamically allocating
the storage associated with each read port. Since it's now doing it
dynamically (it used to be statically allocated), 2 assertions that
were checking the static size were irrelevant and removed.

Furthermore, there was no longer any need to specify a static size
in any port declaration if the port's bandwidth or latency exceeded
the default static size. So, all of these template parameters were
removed to make the code easier to read and understand.