CSN: CSN-asimcore-57
Updating with simcore changes (rev. 2268)

Fix a problem that lead to recursive clearing of the same reference.
When the ref count reached 0 for a pointer being cleared the destructor
was being called before the pointer was cleared. This can lead to
recursive attempts to clear the pointer if the destructor checks the
pointer and tries to clear it. (E.g. Removing an object from a list.)
Changed the code to decrement the ref count, clear the pointer and then
call the destructor if the ref count is 0.