CSN: CSN-asimcore-83
Updating with simcore changes (rev. 2292-2294: Joel Emer)

The commit adds support for finding repository .pack files from a
list of places rather than an single file asim.pack. This packfile
path can contain either individual .pack files or directories, in
which case all of the files (matching the pattern *.pack) will be
used. 

The default packfile path is 
$HOME/.asim/repositories.d
$HOME/.asim/asim.pack
<install-dir>/etc/asim/repositories.d
<install-dir>/etc/asim/asim.pack

That list can be overridden in asim-shell on the command line with
the --repositoryDB switch.

A later phase of this process will involve populating the repositories.d
directory.

Joel

Fixed Makefile bug.