CSN: CSN-asimcore-99
Updating with simcore changes (rev. 2320 & 2321 - Carl Beckmann)

- Multithreaded simulation support:
Added a routine in Clockable class to return the Asim thread a module is
running in, to pass to the clock server.
Fixed a bug in clock server's handling of rate matcher ports.
Added a thread-safe version of the cache manager class.
Added some static params to the multi-chip system controller to enable
multithreaded simulation at various levels of granularity.

- Forgot to add new files from previous commit