CSN: CSN-asimcore-102
Updating with simcore changes (rev. 2323 - Joel Emer)

Added more uniform support for submodels. So in the base
library there is now support for a database of submodels,
i.e., .apm files whose root is not of asim-type "model".
A cache of this database is kept in <AWBLOCAL>/var, like
the module database, so a rehash of that cache is done
with the asim-shell command "rehash models" as well as
will a simple "rehash".

And in apm-edit, the box in the lower left showing
alternative modules lists submodels as alternatives, and
the "refresh" button (the terminology shouldn't be
different) will rehash the submodel cache.

Joel