CSN: CSN-asimcore-104
Updating with simcore changes (rev. 2325 - Brian Slechta)

added support for virtual netbatch. the flow has been tested in MMDC using
regular netbatch. the benchmarks are setup in the local workspace. run areas
and executables are rsync'd to the remote area, the runs are sent to netbach
to run in the remote area, and funally the results are then rsync'd back to the
workspace.

i will now test in AN using virtual netbatch options.