CSN: CSN-asimcore-105
Updating with simcore changes (rev. 2326 - Brian Slechta)

fixes to eliminate uses of undefined variables in the perl modules. when run
on certain machines, the number of warning makes asim tool almost unusable.

fixes to the support for remote virtual netbatch. it "seems" to be working,
but i cannot get the test environment to work. <fingers crossed>