CSN: CSN-asimcore-112
Updating with simcore changes rev. 2336 (Joel Emer)

Added apm-find-replace a graphical program that can be
used to do bulk replacements of a module across a set of
models.Invoke as: apm-find-replace or apm-find-replace --help
for some documentation.

Changes in Model.pm to add a method to intelligently add
a module to a model.

Changes in Inifile.pm to add a method to back up a file.
Backups are in the GNU emacs style of .~<version>~.

Joel