CSN: CSN-asimcore-118
Updating with simcore changes (rev. 2342 - Joel Emer)

apm-find-relace was updated with enhanced filtering capabilities
and the ability to rename models.

The semantics of find_module_{providing,requiring}() in Asim::Model
and Asim::Module have changed to not include searching of the
modules in a submodel. To get the previous behavior one much first
invoke emded_submodles on the model. The only script that actually
relied on the old semantics was hasim-connect, and it has been
updated.

Joel