CSN: CSN-asimcore-120
Updating with simcore changes (rev. 2344 - Michael Powell)

Bugfix to asim-run suggested by adhilton. $bm_name had been converted
to local variable and was being used uninitialized in various places. Fixed
that.

Note: this fix is barely tested and I'm still getting lots of errors from
regression.launcher, but I think this is a better state than it was in.