CSN: CSN-asimcore-121
Updating with simcore changes (rev. 2345 - Brian Slechta)

did a *lot* of cleanup in asim-run. the major changes include the explicit
passing of variables to functions so that we can avoid all the problems of
using the global variables. this seems to fix all of the remaining problems
that i am aware of. both local build and batch build modes seem to work great
now. <fingers_crossed> also, i fixed the "use of unitialized variables" in
PutEnv. the rest of the changes are just cleanup.

added WARN and WARNX to mesg.h. these macros are like ASSERT but they do not
kill the program.