CSN: CSN-asimcore-128
Updating with simcore changes (rev. 2353 - Joel Emer)

Added an 'Admin' tab to awb. This tab allows one to create
and switch workspaces, checkout and update packages. Some
additional functionaly is disabled pending more support
for interactive responses in a graphical environment in
asim-shell.

Partial fix for awb sometimes not exiting completely and
the 'Kill' button not seeming to function in the run log.
Problems still arise when the subprocess does not respond
to a SIGPIPE.