CSN: CSN-asimcore-139
Updating with simcore changes (rev. 2363 - Brian Slechta)

fixed create-benchmark-file to work with NLIT traces using multi.cfx. this
script should now work properly with all traces currently in the Master.tlist.

created a special rsync function in asim-run that can be modified to try on
errors. it still does not rety anything, but the framework is there.

added rsync error checking and rety to asim-batch. if rsync has a non-zero
exit code, it will retry a bunch of times.