CSN: CSN-asimcore-168
Updating with simcore changes (rev. 2407 - Abhishek Bhattacharjee)

Changes made to accomodate integrated power macro calls in ports for
models using the power infrastructure. In port.h, the access type of
the private members in ReadPort, WritePort, ReadSkidPort,
WriteSkidPort etc. are changed to protected. This is so that the
PowerReadPort, PowerWritePort etc. classes in the asim-power package
derived from the classes in port.h have access to those members as
well.
null_power_model.h has been modified to include the null model for the
power ports.