CSN: CSN-asimcore-179
Updating with simcore changes (rev. 2429 - Carl Beckmann)

- Support for AWB-friendly generic .cfx files.
Now you can create generic .cfx scripts that get their benchmark name
from what follows .cfx in the directory path, but there is now an additional
Directory() command in GenCFG::Auto that generic .cfx scripts can use to
display a directory of benchmarks that gets displayed in AWB.
(Got this partially from the MT branch, with some additional changes).

Modified the multi-tlist.cfx script so it can be used with AWB. If you make
a copy of (or soft link to) multi-tlist.cfx, then in AWB it will find any
.tlist files in the directory hierarchy below it, and it will set up a
multicore benchmark run to run all the listed traces in parallel on the model.
It works with pure-Asim style multicore models.