CSN: CSN-asimcore-191
Updating with simcore changes (rev. 2448 - Carl Beckmann)

- The classic controller's "-cfg <filename>" command line option
now accepts relative pathnames for <filename>. The path can be
relative to either the current working directory, or in the case
of recursive -cfg files, relative to the directory in which the
enclosing config file resides. In the case of a conflict, a warning
message is issued and the one relative to the enclosing config file
is used.