CSN: CSN-asimcore-194
Updating with simcore changes (rev. 2450 - Brian Slechta)

enhanced the heartbeat to include MacroInsts in addition to Insts.

add "-pm" for macro instruction heartbeat.

enhanced the ASIM_ITEM with thread id support to be used with PTV event
records.

added option to asim-batch to allow the run logs to be saved so that we can
get at simulator output (like heartbeat for ipc analysis)