CSN: CSN-asimcore-230
Updating with simcore changes (rev. 2554 - Joel Emer)

In asim-shell and awb:
Generalized many of the package commands to take 'all' or a list of
package names.


Just in asim-shell:
Added --source switch to 'install package' to install a public version
of a package. Underlying support for this was aded to Asim perl library.


In Asim perl library and apm-edit:
Use Asim::invoke_editor(), Asim::get_editor() and Asim::get_tmpdir()


In Asim.pm
Centralized the definitions and use of the EDITOR and TMPDIR environment
variables and added use of ASIMEDITOR and ASIMTMPDIR by adding the
following functions:

Asim::invoke_editor() - to invoke an editor on a file. Editor is
obtained from Asim::get_editor()

Asim::get_editor() - get the GUI-style editor command defined by
ASIMEDITOR or EDITOR

Asim::get_tmpdir() - to get the name of the temporary directory
defined by ASIMTMPDIR or TMPDIR.