CSN: CSN-asimcore-240
Enhanced the search for rcfile options to include the file
<workspace>awb.config. So now program's such as regression.*
and asim-run will look for their options first in awb.config,
then in ~/.asim/asimrc and then in <installdir>/etc/asim/asimrc.

All the scripts that use options have been enchanced with an
--options switch that lists the options that the program uses,
and their current values.

All of this means that one can set options on a per workspace
basis, so if, for example, one has a workspace for working on
hasim, then they can set up default regressions for that workspace
to regress the hasim packages by adding the following to their
awb.config file for that workspace.

awb.config:
.
.
.
[regression.launcher]
PACKAGES=hasim,hasim-mips
.
.
.

Finally, the change to inifile.cpp in libawb is to allow dot (.)
in a inifile group name, and must be installed system-wide before
the above funcionality will function correctly.