CSN: CSN-asimcore-246
Updating with simcore changes (rev. 2572 - Joel Emer)

Refactored the batch code in asim-run so that the batch subsystem code
is in separate perl modules. The idea is that there are separate
modules for supporting netbatch, netbatch-remote, local and eventually
other batch subsystems. Each module needs to support the batch
operations that we supported directly in asim-rum in a batch system
specific way, e.g., check size of a directory, check number of
threads, schedule to reduce number of jobs in-flight, and submit the
job. The data sloshing for the remote runs is still done manually in
asim-run and asim-batch.