CSN: CSN-asimcore-248
Updating with simcore changes (rev. 2574 - Joel Emer)

Changes to many of the asim-shell command on sets of packages so
that they continue to process all the packages in the set even in
the event of various problems on an individual package. A summary
of the failures is then printed at the end of the command.