CSN: CSN-asimcore-250
Updating with simcore changes (rev. 2577 - Joel Emer)

In asim-shell, have 'update package' build using the common
configure_and_build_packages subroutine. This routine
includes properly ordering builds by package priority.

In Fork.pm, open child's STDOUT/STDERR logfile in a way
that preserves the redirection in subprocesses of the child.
Also, added a convenience predicate argument in the
controlled_fork and controlled_exit routines.

In asim-run, use the new interface to controlled_fork and
controlled_exit and properly reset the $exit_code value
on entry to a child, so that a pending error code in the
parent isn't propagated to all the children.

In Inifile.pm, fix a minor bug in the perldoc directives.