CSN: CSN-asimcore-272
Updating with simcore changes (rev. 2632 - Steve Raasch)

asim-run: add output reminding the user of which pool jobs are being
submitted to

asim-batch: now creates 'run.info' file in the benchmark's RUN
directory that describes how the simulation was invoked. This file
duplicates the content of 'run.error.*' except that it is created for
all runs, not just those that fail