CSN: CSN-asimcore-274
Updating with simcore changes (rev. 2634 - Eric Borch)

- added ability to specify a directory in the BUILD directory to hold
files that are read in at run time.  This is needed for remote
netbatch environments.  With this feature, these run time files
will be copied to the remote netbatch location by asim-run.  The directory
can be specified in asimrc with RUN_TIME_DIR= or on the command line with
--run_time_dir.  Right now this ONLY allows directories that are in the
BUILD directory.

- there's also a change to not rsync anything to a remote netbatch
location if the --nobuild option is specified.

Eric