CSN: CSN-awb-322
Added implementation of the notion of a 'template' module, which 
can be used to hold a skeleton implementation for modules of the
given awb-type and hold notes relevant to all modules of that awb-type.

A template module is indicated by a --template flag on the %provides 
directive in an .awb file

Template modules show up as alternatives in apm-edit, but cannot be inserted
into the model.

Also put in a fix for the periodic crash experienced by apm-edit from 
what appears to be a perlqt bug on certain ListView method calls. Fix was
to put an eval{} expression around the suspect method calls.