CSN: CSN-awb-337
For 'global' awb parameters defined in a module within a submodel that
itself is included in an outer submodel, we now allow changing such
parameters at the level of the model that includes the outer submodel.
In other words, 'global' awb parameters are always visible regardless of
whether they are buried in multiple levels of submodels.