CSN: CSN-awb-369
Added a -completioncmd flag to asim-run.  This flag will run a
specified command at the end of a regression run.  In the case of
condor, this will cause a DAG of jobs to be specified.  For local
runs, the command will run serially at the end of the last job in the
regression.  I made no attempt at porting to Netbatch.  I use the
feature for nightly regressions, when I want plot.nightly to run at
the end of my regression, rather than guessing when I should run this
script.