CSN: CSN-awb-376
Initial support for building debian packages for awb.

To directly build the source and then the binary package
type the following:

     % cd <root-of-awb-package>
     % debuild -rsudo -S
     % debuild -rsudo

Some things that still need to be done are complete 
dependency specifications and proper placement of 
configuration files into /etc. I'm sure that's not all.

Building with pbuilder has not been tested.

Joel