CSN: CSN-awb-377
More .deb support. 

When installing using a .deb file, the conffiles, i.e., the
files in etc/asim, are now put into /etc/asim instead of
PREFIX/etc/asim. Both asimstarter and Asim.pm now check in
/etc/asim first before using the default, PREFIX/etc/asim.

Note: if we eventually make our nomenclature consistent and
use 'awb' instead of 'asim' those files will need to change. Or
maybe configure.in should handle this...

The name of the system-wide asimrc file in the repository
was changed from asimrc.template to asimrc to make things
easier for the .deb install procedure. It is still just called
asimrc in etc/.

Moved the simple template workspace from etc/asim/ws to 
share/asim/ws, since it is not really a configuration file
and it was complicating the debinan installs. That change
reguired changes in configure.in, and various Makefiles, so
autoconf made lots of gratuituous changes.