CSN: CSN-awb-379
Added a make target of install-package, which installs
the module pool (and base/) in $codedir, which is 
currently PREFIX/share/asim/<package>/<version>. The
debian install does that install so that one can
have the awb module pool even without checking out
the package.

Make a few other improvements for .deb creation.