CSN: CSN-awb-382
Added symlink from awb to awb2 in libexec, so awb launches 
properly when using the centrally installed version.

Fixed perl version check in asimstarter.

Added cvs to list of required packages for the .deb package

Moved check for the environment variable CSH_RSH into the 
usage of CVS to avoid the annoying messages about it when 
one isn't using CVS.

Did an autoreconf to upgrade to a new version of autotools.