CSN: CSN-awb-383
Added the Leap (Logic-based Environment for Application
Programming) .pack file, which defines the new Leap 
repository and the associated Leap platform repository.
The HAsim .pack file now defines only modeling-related 
repositories.

Added support for a awb model type of 'Leap', which use
a set of leap-* commands, such as leap-configure, in 
place of the misnamed hasim-* commands.

Updated asimstarter to understand the leap-* commands,
which means that a new install of awb will be required
before one can use the commands from the (soon to be 
populated) Leap package.

Note that .apm files with model type 'HAsim' will currently
continue to use the hasim-* commands in the HAsim package. 
Eventually, however, a checkout of the Leap package will
be required to build.