CSN: CSN-awb-388
A variety of improvements to documentation and to distinguish 
awb from asim.

Changed name of various asim-<name> programs to be called 
awb-<name>.  Maintaining backwards compatibility via symbolic 
links and having the programs look for rc options first in group 
[awb-<name>] and then in [asim-<name>]. After all your workspaces 
are updated to this version of the awb pacakge you should change 
the asim- groups to be awb- in ~/.asim/asimrc.

Added creation of man files from the perl documentation for
a number of programs. In the process, fixed a number of perlpod 
syntax errors.

Converted perl builds to type 'vendor', which puts files (in
particular man page) where we want them, and doesn't generate 
the superflous podlocal files, which now don't need to be deleted 
during install. In the process, fixed a bunch of perl MANIFEST 
files.

Bug fix on locating <INSTALLDIR>/etc/asim/asimrc, when awb is 
installed using the .deb file.