CSN: CSN-awb-401
Added support for creating a new repository for a package.
Currently only svn repositories are supported. 

The procedure for creating a new repository is:

1) Create an empty repository on your sever

2) Run the following commands

   % awb-shell create package <pkgname>
   ...answer questions...
   % awb-shell create repository <pkgname>
   ...answer questions...
   % awb-shell checkout package <pkgname>
   ...answer questions...

A .pack file is also created, which can be
installed globally, e.g., by adding it to
the awb package.  This .pack file is placed in:

   ~/.asim/repositories.d/<pkgname>.pack