CSN: CSN-awb-449
Add repository information for release v11.04 of awb,
asim-core, leap and platforms packages.

Make debian package depend on both the current and the
previous release of package. Thus, an upgrade will not
automatically wipe out previous version of the package 
in /usr/share/asim/awb/. It would be better if a new
install just left previous versions alone, but I couldn't
get that to work.