CSN: CSN-awb-453
Updated pack files for following package name changes:

           platforms     -> leap-platforms
           platforms-acp -> leap-platforms-acp


The repostory  names were changed as well as the locations
of the repositories on the svn server and the checkout 
targetdirectory.

Symbolic links from old names to new names were placed
on the svn server so that old checkouts will continue
to work.