CSN: CSN-awb-477
Improved behavior of parameters of submodels. New behavior is
related to the default value of a submodel parameter as seen by
the parent model, i.e., the value of 'global' parameter in the 
submodel. Previously, that value was taken only from the module's 
.awb file. Now that default value may be taken from an override of 
the parameter value in the submodel. This saves a lot of writing 
of parameter values in the parent model's .apm file, and allows 
the value of a submodel parameter to be changed in all model's 
that use the submodel simply by changing the value in the submodel.