CSN: CSN-awb-490
Fixed regression in behavior of parameters introduced in 476.
Now the defaults of global parameters of a submodel are not 
modified when overridden by a parent model, so that saves 
of the submodel properly save the overwridden values. In
addition, only parameters whose values differ from the
submodels default are saved in the .apm file.