CSN: CSN-awb-502
Tab completion in awb-shell now works for .cfx files.

The awb-shell 'source' command now looks to see if
the input file is executable and if it is invokes
that file (passing in the additional arguments on the
command line) and executes the series of commands 
printed to STDOUT by that executable. Note the arguments
are not treated specially so fancy quoting or substitutions
will probably not work.