Added initial support for Perforce (p4) repositories.
The following commands are supported;
    - checkout
    - update
    - delete

The 'checkout' command creates a p4 config file (P4CONFIG) and a p4 client
and the 'delete' command will delete the client (and delete the local files).
But there is only minimal error checking, so one needs to manually make sure
that the local directory is in a clean state.